+++ src/rating/index.tsx
@@ -1,20 +1,2 @@
-import styled from 'styled-components'
-
-import { color, space } from '../_utils/branding'
-import { Rating } from './Rating'
-
-const StyledRating = styled(Rating)`
-  & {
-    display: flex;
-    align-items: center;
-  }
-
-  & span {
-    color: ${color.lightMidnightGreen};
-    margin-left: ${space.s};
-  }
-`
-
-export { RatingProps } from './Rating'
-export { StyledRating as Rating }
-export default StyledRating
+export { Rating, RatingProps } from './Rating'
+export { Rating as default } from './Rating'