+++ src/emptyState/EmptyState.tsx
@@ -7,14 +7,23 @@ import { StyledEmptyState } from './EmptyState.style'
 export type EmptyStateProps = Readonly<{
   className?: string
   image: string
-  text: string
+  text: string | JSX.Element
   button?: JSX.Element
 }>
 
-export const EmptyState = ({ className, image, text, button }: EmptyStateProps) => (
-  <StyledEmptyState className={cc(['kirk-empty-state', className])}>
-    <img src={image} alt="" />
-    <Title>{text}</Title>
-    {button}
-  </StyledEmptyState>
-)
+export const EmptyState = ({ className, image, text, button }: EmptyStateProps) => {
+  let title
+  if (typeof text === 'string') {
+    title = <Title>{text}</Title>
+  } else {
+    title = text
+  }
+
+  return (
+    <StyledEmptyState className={cc(['kirk-empty-state', className])}>
+      <img src={image} alt="" />
+      {title}
+      {button}
+    </StyledEmptyState>
+  )
+}