+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Unreleased
 
+- **[UPDATE]** Update `EmptyState` component type for `text` prop (`string | JSX.Element`)
 - **[UPDATE]** Migrate `MediaSection` to new mdx story format
-[...]
+  [...]
 
 # v40.0.3 (03/09/2020)