+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - **[UPDATE]** Update `EmptyState` component type for `text` prop (`string | JSX.Element`)
 - **[UPDATE]** Migrate `MediaSection` to new mdx story format
-  [...]
+- **[FIX]** Fixed wrong spacing between icon & text in `Button`
 
 # v40.0.3 (03/09/2020)