+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+[...]
+
+# v40.1.0 (09/09/2020)
+
 - **[UPDATE]** Update `EmptyState` component type for `text` prop (`string | JSX.Element`)
 - **[UPDATE]** Migrate `MediaSection` to new mdx story format
 - **[FIX]** Fixed wrong spacing between icon & text in `Button`